+++ ChangeLog.rst
@@ -1,6 +1,15 @@
 ChangeLog
 =========
 
+0.1.1
+-----
+
+* `ViewCounter.counter` was renamed to `count`.
+* `DownloadCounter.cnt` was renamed to `count`.
+* `Referer.cnt` was renamed to `count`.
+* `DownloadCounter.redir_url` was renamed to `url`.
+* `DownloadCounter` was renamed to `RedirCounter`.
+
 0.1.0
 -----